refactor(kv-tokens): consume generated tokens in tailwind config and build scripts
